Cycles : Move `convertPrimitiveVariable()` to GeometryAlgo
It is used by all the GeometryAlgo converters, and doesn't really warrant being in its own namespace.
