+++ src/GafferCycles/IECoreCyclesPreview/GeometryAlgo.cpp
@@ -38,6 +38,8 @@
 
 #include "IECoreScene/PrimitiveVariable.h"
 
+#include "IECore/SimpleTypedData.h"
+
 IECORE_PUSH_DEFAULT_VISIBILITY
 // Cycles (for ustring)
 #include "util/param.h"
@@ -88,6 +90,95 @@ Registry &registry()
 {
 	static Registry r;
 	return r;
+}
+
+ccl::TypeDesc typeFromGeometricDataInterpretation( IECore::GeometricData::Interpretation dataType )
+{
+	switch( dataType )
+	{
+		case GeometricData::Numeric :
+			return ccl::TypeDesc::TypeVector;
+		case GeometricData::Point :
+			return ccl::TypeDesc::TypePoint;
+		case GeometricData::Normal :
+			return ccl::TypeDesc::TypeNormal;
+		case GeometricData::Vector :
+			return ccl::TypeDesc::TypeVector;
+		case GeometricData::Color :
+			return ccl::TypeDesc::TypeColor;
+		case GeometricData::UV :
+			return ccl::TypeDesc::TypePoint;
+		default :
+			return ccl::TypeDesc::TypeVector;
+	}
+}
+
+template<typename T>
+size_t dataSize( const TypedData<std::vector<T>> *data )
+{
+	return data->readable().size();
+}
+
+template<typename T>
+size_t dataSize( const TypedData<T> *data )
+{
+	return 1;
+}
+
+template<typename T>
+ccl::Attribute *convertTypedPrimitiveVariable( const std::string &name, const PrimitiveVariable &primitiveVariable, ccl::AttributeSet &attributes, ccl::TypeDesc typeDesc, ccl::AttributeElement attributeElement )
+{
+	// Get the data to convert, expanding indexed data if necessary, since Cycles doesn't support it
+	// natively.
+
+	const T *data;
+	typename T::Ptr expandedData;
+	if( primitiveVariable.indices )
+	{
+		expandedData = boost::static_pointer_cast<T>( primitiveVariable.expandedData() );
+		data = expandedData.get();
+	}
+	else
+	{
+		data = static_cast<const T *>( primitiveVariable.data.get() );
+	}
+
+	// Create attribute. Cycles will allocate a buffer based on `attributeElement` and the information
+	// `attributes.geometry` contains.
+
+	ccl::Attribute *attribute = attributes.add( ccl::ustring( name.c_str() ), typeDesc, attributeElement );
+
+	// Sanity check the size of the buffer, so we don't run off the end when copying our data into it.
+
+	const size_t expectedSize = attribute->element_size( attributes.geometry, attributes.prim );
+	if( dataSize( data ) != expectedSize )
+	{
+		msg(
+			Msg::Warning, "IECoreCyles::GeometryAlgo::convertPrimitiveVariable",
+			boost::format( "Primitive variable \"%1%\" has size %2% but Cycles expected size %3%." )
+				% name % dataSize( data ) % expectedSize
+		);
+		return nullptr;
+	}
+
+	// Copy data into buffer.
+
+	if constexpr( std::is_same_v<T, V3fVectorData> || std::is_same_v<T, Color3fVectorData> )
+	{
+		// Special case for arrays of `float3`, where each element actually contains 4 floats for alignment purposes.
+		ccl::float3 *f3 = attribute->data_float3();
+		for( const auto &v : data->readable() )
+		{
+			*f3++ = ccl::make_float3( v.x, v.y, v.z );
+		}
+	}
+	else
+	{
+		// All other cases, (including int to float conversion) are a simple element-by-element copy.
+		std::copy( data->baseReadable(), data->baseReadable() + data->baseSize(), attribute->data_float() );
+	}
+
+	return attribute;
 }
 
 } // namespace
@@ -149,6 +240,107 @@ ccl::Geometry *convert( const std::vector<const IECore::Object *> &samples, cons
 void registerConverter( IECore::TypeId fromType, Converter converter, MotionConverter motionConverter )
 {
 	registry()[fromType] = { converter, motionConverter };
+}
+
+void convertPrimitiveVariable( const std::string &name, const IECoreScene::PrimitiveVariable &primitiveVariable, ccl::AttributeSet &attributes, ccl::AttributeElement attributeElement )
+{
+	ccl::Attribute *attr = nullptr;
+	switch( primitiveVariable.data->typeId() )
+	{
+		// Simple int-based data. Cycles doesn't support int attributes, so we promote to the equivalent float types.
+
+		case IntDataTypeId :
+			attr = convertTypedPrimitiveVariable<IntData>( name, primitiveVariable, attributes, ccl::TypeDesc::TypeFloat, attributeElement );
+			break;
+		case V2iDataTypeId :
+			attr = convertTypedPrimitiveVariable<V2iData>( name, primitiveVariable, attributes, ccl::TypeFloat2, attributeElement );
+			break;
+		case V3iDataTypeId :
+			attr = convertTypedPrimitiveVariable<V3iData>(
+				name, primitiveVariable, attributes,
+				typeFromGeometricDataInterpretation(
+					static_cast<const V3iData *>( primitiveVariable.data.get() )->getInterpretation()
+				),
+				attributeElement
+			);
+			break;
+
+		// Vectors of int-based data. Cycles doesn't support int attributes, so we promote to the equivalent float types.
+
+		case IntVectorDataTypeId :
+			attr = convertTypedPrimitiveVariable<IntVectorData>( name, primitiveVariable, attributes, ccl::TypeDesc::TypeFloat, attributeElement );
+			break;
+		case V2iVectorDataTypeId :
+			attr = convertTypedPrimitiveVariable<V2iVectorData>( name, primitiveVariable, attributes, ccl::TypeFloat2, attributeElement );
+			break;
+		case V3iVectorDataTypeId :
+			attr = convertTypedPrimitiveVariable<V3iVectorData>(
+				name, primitiveVariable, attributes,
+				typeFromGeometricDataInterpretation(
+					static_cast<const V3iVectorData *>( primitiveVariable.data.get() )->getInterpretation()
+				),
+				attributeElement
+			);
+			break;
+
+		// Simple float-based data.
+
+		case FloatDataTypeId :
+			attr = convertTypedPrimitiveVariable<FloatData>( name, primitiveVariable, attributes, ccl::TypeDesc::TypeFloat, attributeElement );
+			break;
+		case V2fDataTypeId :
+			attr = convertTypedPrimitiveVariable<V2fData>( name, primitiveVariable, attributes, ccl::TypeFloat2, attributeElement );
+			break;
+		case V3fDataTypeId :
+			attr = convertTypedPrimitiveVariable<V3fData>(
+				name, primitiveVariable, attributes,
+				typeFromGeometricDataInterpretation(
+					static_cast<const V3fData *>( primitiveVariable.data.get() )->getInterpretation()
+				),
+				attributeElement
+			);
+			break;
+		case Color3fDataTypeId :
+			attr = convertTypedPrimitiveVariable<Color3fData>( name, primitiveVariable, attributes, ccl::TypeDesc::TypeColor, attributeElement );
+			break;
+
+		// Vectors of float-based data.
+
+		case FloatVectorDataTypeId :
+			attr = convertTypedPrimitiveVariable<FloatVectorData>( name, primitiveVariable, attributes, ccl::TypeDesc::TypeFloat, attributeElement );
+			break;
+		case V2fVectorDataTypeId :
+			attr = convertTypedPrimitiveVariable<V2fVectorData>( name, primitiveVariable, attributes, ccl::TypeFloat2, attributeElement );
+			break;
+		case V3fVectorDataTypeId :
+			attr = convertTypedPrimitiveVariable<V3fVectorData>(
+				name, primitiveVariable, attributes,
+				typeFromGeometricDataInterpretation(
+					static_cast<const V3fVectorData *>( primitiveVariable.data.get() )->getInterpretation()
+				),
+				attributeElement
+			);
+			break;
+		case Color3fVectorDataTypeId :
+			attr = convertTypedPrimitiveVariable<Color3fVectorData>( name, primitiveVariable, attributes, ccl::TypeDesc::TypeColor, attributeElement );
+			break;
+		default :
+			msg( Msg::Warning, "IECoreCyles::GeometryAlgo::convertPrimitiveVariable", boost::format( "Primitive variable \"%s\" has unsupported type \"%s\"." ) % name % primitiveVariable.data->typeName() );
+			return;
+	};
+
+	// Tag as a standard attribute if possible. Note that we don't use `AttributeSet::add( AttributeStandard )`
+	// because that crashes for certain combinations of geometry type and attribute. But some of those "crashy"
+	// combinations are useful - see `RendererTest.testPointsWithNormals` for an example.
+	//
+	/// \todo Support more standard attributes here. Maybe then the geometry converters could
+	/// use `convertPrimitiveVariable()` for most data, instead of having
+	/// custom code paths for `P`, `uv` etc?
+
+	if( name == "N" && attr->element == ccl::ATTR_ELEMENT_VERTEX && attr->type == ccl::TypeDesc::TypeNormal )
+	{
+		attr->std = ccl::ATTR_STD_VERTEX_NORMAL;
+	}
 }
 
 } // namespace GeometryAlgo
